The mean, median and focal value liones are working. Need better colors and a solution to the legend. Ideally a hybrid between the subplot solutio and the current notebook based design.
