+++ client/src/components/ProTx/components/charts/ObservedFeaturesChart.js
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
-// import ObservedFeaturesPlot from './ObservedFeaturesPlot';
-// import ObservedFeaturesPlotHardcoded from './ObservedFeaturesPlotHardcoded';
-import ObservedFeaturesPlotRefactor from './ObservedFeaturesPlotRefactor';
+// import ObservedFeaturesPlot from './ObservedFeaturesPlotOld';
+// import ObservedFeaturesPlot from './ObservedFeaturesPlotHardcoded';
+import ObservedFeaturesPlot from './ObservedFeaturesPlot';
 import ChartInstructions from './ChartInstructions';
 import './ObservedFeaturesChart.css';
 
@@ -46,10 +46,7 @@ function ObservedFeaturesChart({
   if (selectedGeographicFeature && observedFeature) {
     return (
       <div className="observed-features-report">
-        {/* <ObservedFeaturesPlot */}
-        {/* <ObservedFeaturesPlotHardcoded */}
-        {/* <ObservedFeaturesPlotRefactor */}
-        <ObservedFeaturesPlotRefactor
+        <ObservedFeaturesPlot
           mapType={mapType}
           geography={geography}
           observedFeature={observedFeature}
